SingleFile
I have collapsed the three files into one file. It is called get_where_used.php. The sql statements in the apiUtility have been updated as well.
